Remove all AMD/Browser/minified bits
If users want to build things, they should use Webpack, or Browserify,
or one of the many other options.  There's no way to please everyone, so
it seems the best approach is to just be like all the other modules that
do not do this stuff.

Separation of concerns.

Fix #120