fix: allow `.` to prefix prerelease in loose mode
BREAKING CHANGE: prerelease tag requires a `-` if it doesn't start with
a non-numeric value.  If it does start with a non-numeric value the `-`
is optional and a `.` can be used instead.  This means that `1.2.34.5`
is no longer a valid version in loose mode (previously it would split
into `1.2.3-4.5` and `1.2.3.4` is invalid (previously it would split
into `1.2.3-4`.

Allowing 1.2.3.4 is confusing and weird.

However, 1.2.3.beta should be allowed.