Handle X ranges properly in includePrelease mode
Append a `-0` prerelease tag on the parsed comparators when in
includePrerelease mode.  Otherwise, 2.x will be satisfied by 3.0.0-beta,
since it's in `>=2.0.0 <3.0.0`.  However, `2.0.0-beta` will _not_
satisfy this range.

By appending the `-0`, the version set is pinned to the appropriate
boundary.

Note that doing this in non-prerelease-including mode would be a
mistake!  Appending the `-0` to the comparator means that prereleases in
the tuple are allowed, so `2.0.0-beta` would match `2.x` in
non-prerelease-including mode as well.  However, if we're including
prereleases, then their inclusion is intentional.

Fix: https://github.com/npm/node-semver/issues/282