fix: Handle prereleases properly in 'X - Y' ranges
If includePrerelease is set, then you'd expect 1.0.0-pre and 2.0.0-pre
to be included in the range '1.0.0 - 2.0.0'.  You would NOT expect that
3.0.0-pre would be included in that range.

To accomplish this, the bounds of the range have to be set differently,
using the '-0' minimum prerelease version identifier, to include and
exclude prereleases appropriately when includePrerelease is set.