Exclude prereleases in X-ranges with < comparators
This one didn't cause any bugs, but it's more for consistency and
future-proofing.

When a range like '<=1' is translated to 'anything less than 2.0.0', it
should be expressed as '<2.0.0-0' to prevent prerelease 2.0.0 versions
as well.

It didn't cause any problems with intersection or range satisfying
because it would only treat 2.0.0-pre as satisfying <2.0.0 when
includePrerelease was set, and if includePrerelease was set, then it
would add the -0 to the less-than comparator.

Nevertheless, all other areas where we use a <X.Y.0 range, we add the -0
prerelease in order to force the boundary of the set to exclude
prereleases on that version, which is good defense in depth to prevent
other errors with intersection and subsets seen previously.