Memoize range parsing
This yields about a 10-20% speed improvement to a lot of common npm
operations that spend most of their CPU time parsing a relatively small
set of dependency ranges.

It does mean that one `range.set` may contain the same Comparator
objects as another range with the same sub-range.  That was technically
already possible, due to the `Comparator.ANY` magic range, but it did
require a bit of tweaking in the `subset` function, which is also now
a bit faster when calculating subsets for ranges that are partly or
entirely identical.